Add documentation for the views
The documentation is inline in the form of comments that describe the
columns in the views, the tables from which we pull data and details on
why/how we're doing things.

There are a couple of `TODO` notes with questions about what the
underlying data means which will allow us to confirm
suspicions/assumptions.