Update School Experience statuses
The withdrawn status was mistakenly used when a candidate did not attend a placement (it should be used when a school cancels a placement request that has not yet been accepted).

This has now been fixed in the CRM and School Experience app. The `withdrawn` status has been updated to `did not attend`, and a new status has been added for `withdrawn`.